fix cs + move getExportOptions to RewrittenInputTableOptions
